+++ nacc_attribute_deriver/attributes/nacc/modules/uds/uds_namespace.py
@@ -1,9 +1,12 @@
 """Class to define UDS-specific attributes."""
 
 from datetime import date, datetime
-from typing import Optional
+from typing import Any, Optional
 
-from nacc_attribute_deriver.attributes.base.namespace import FormNamespace
+from nacc_attribute_deriver.attributes.base.namespace import (
+    FormNamespace,
+    SubjectDerivedNamespace,
+)
 from nacc_attribute_deriver.schema.errors import MissingRequiredError
 from nacc_attribute_deriver.symbol_table import SymbolTable
 
@@ -13,9 +16,42 @@ class UDSNamespace(FormNamespace):
         """Check that this is a UDS form."""
         super().__init__(table)
 
-        module = self.get_value("module")
+        self.__subject_derived = SubjectDerivedNamespace(table)
+
+        module = super().get_value("module")
         if not module or module.upper() != "UDS":
             raise MissingRequiredError("Current file is not an UDS form")
+
+    def is_followup(self) -> bool:
+        """Returns whether or not this is a follow-up form."""
+        return super().get_value("packet") == "F"
+
+    def check_default(self, attribute: str, default: Optional[Any] = None) -> Any:
+        """Check for the default by:
+
+            1. If a follow-up packet and result is None/777, check
+                subject.info.derived.<attribute>
+            2. If still None, then return the specified default
+
+        NOTE: The 777 will often NOT be attached to the derived attribute,
+            but one of the sources it looks at. It's here to cover the case,
+            but in general it is expected a 777 value should fall to the
+            default case that calls this method for that specific rule.
+
+        Args:
+            attribute: the target attribute name
+            default: the default value
+        Returns:
+            the value for the attribute in the table
+        """
+        if not self.is_followup():
+            return default
+
+        result = super().get_value(attribute)
+        if result in [None, 777, "777"]:
+            result = self.__subject_derived.get_value(attribute, None)
+
+        return default if result is None else result
 
     def generate_uds_dob(self) -> Optional[date]:
         """Creates UDS DOB, which is used to calculate ages."""